ci: build arm binaries
I have split build.sh in two parts. The first sets up and destroys the
chroot, the other runs the actual builds command inside of the Alpine
environment.

This makes it possible to run chroot_build.sh on CI's native
architecture so that it handles creation of the native chroot, and run
only build.sh in the run-on-arch action, that creates an Alpine
container itelf.

I also substituted version numbers with "*" when possible, so that
updating versions of build binaries is easier (you'll only have to
change the download URL)